requirements.txt
discord.py
PyYAML
openai
requests